[Notifications] Allow selecting text with mouse
This was lost by the introduction of the default action as the MouseArea steals events
from the TextField below.
To address this, selecting text and click handling has been implemented manually.
Furthermore, the mouse cursor is changed to an IBeam cursor and turns into a hand when
above hyperlinks whose URL can now also be copied separately from its text in the
context menu.

BUG: 386118
FIXED-IN: 5.12.0

Differential Revision: https://phabricator.kde.org/D8446