Fix system tray settings if placed on desktop
Summary:
The DesktopView replaces the default ConfigView for itself with a
specialised ContainmentConfigView. It's only loaded by the desktop, and
is more of a Desktop-ConfigView than generic containment.

The system tray applet doesn't show it's own config, but instead shows a
config of its internal system tray containment by overriding the
configure action.

This causes the desktop view to mistakenly shows it's specialised
desktop config view when the system tray is configured.

This patch adds an explicit check on the containment type.

Test Plan:
Clicked on sys tray settings on desktop
Clicked on normal desktop settings
Clicked on sys tray settings on panel

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9526