Don't show "Edit Application..." action for non-app services.
Summary:
When the code was refactored to show this action also for runner
matches, proper care wasn't taken to test for this, so the action
was shown for document matches as well. It's not safe to use for
docs currently (which should probably be fixed in
KPropertiesDialog) and at any rate isn't meant to be a generic
editing feature as per the action name.

BUG:388340

Reviewers: #plasma, broulik, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9612