[weather] Dump old code for day/night calculation by time dataengine
Sync querying of the time dataengine is no longer possible,
also is the calculation of the elevation not depending on other data
then those passed in, so this is not really data hold by the time
data engine. Instead one would calculate the values needed on the fly.

Yet to solve: how to make direct use of the solarsystem code currently
part of the time dataengine only.