[Panel View] Check if it's the same applet we're configuring before just hiding current view
This fixes opening panel controller while an applet config dialog was open.
It would find that we want to show config for a panel but find a config is already open, so it would just close it,
even though it was not the panel controller (which should be toggleable) but another applet config.
While at it restructure if statements a bit.

Differential Revision: https://phabricator.kde.org/D9799