Further improve quality of images in notifications
BUG: 385097

This revision further improves the quality of images in notifications
that was addressed in revision D9993. When I first looked into this, I
tried setting ScaleType on the PreviewJob but it had no effect. I found
that the preview images were being scaled to the small vertical size
that was requested, and this resulted in my first revision. I should
have tried setting ScaleType in addition.

Notification with unpatched libnotificationshelperplugin.so

{F5677396}

Notification with patch applied

{F5677397}

Test Plan:
Take screenshot with Spectacle with unpatched
libnotificationshelperplugin.so  installed and observe the blurry
preview image in the notification.

Install the patched libnotificationshelperplugin.so and restart
plasmashell.

Take another screenshot with Spectacle and observe the much improved
quality of the preview image in the notification.

Reviewers: #plasma_workspaces, #plasma, broulik, ngraham

Subscribers: #plasma, #plasma_workspaces, plasma-devel, broulik, ngraham

Tags: #plasma, #spectacle

Differential Revision: https://phabricator.kde.org/D10146