Fix dupe handling in requestAddLauncherToActivities; improve unit test
Summary:
The business logic was using the resolved URL in lookups in
internal data structures, instead of the internal key it
previously established equivalence to. This could lead to
junk in internal maps and unnecessarily emitted model changes.

This also improves the unit test not to require apps to be
installed, fixing the CI failure over missing Dolphin.

Reviewers: #plasma, kossebau

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10253