Improve appearance of the logout dialog on wayland
Summary:
The logout dialog has a taskbar entry and appears like a regular window on
wayland. This is because Qt does not support fullscreen windows with xdg_shell
as of 5.9.4. wl-shell works, so force it for the time being.

Test Plan: Ran ksmserver-logout-greeter on X11 and wayland, looks identical now.

Reviewers: #plasma, graesslin, davidedmundson, bshah

Reviewed By: #plasma, graesslin, bshah

Subscribers: ngraham, bshah, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10425