dont initialize app menu index in every hiding
Summary:
-- hiding a menu shouldnt initialize the currentIndex to -1
at all cases. When a secondary menu is already shown then
the currentIndex should be only the index of the new menu.
This improves also the qml bindings because currentIndex
doesnt change all the time with no reason.

Test Plan:
added debug messages in qml side to check the value of
currentIndex

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10695