Fix the userswitcher when using the mouse for switching
Summary:
Clicking on a button makes Qt grab the pointer, which means the screen can't be
locked. As we wait for the screen to be locked before switching, this never
happens.

BUG: 391007

Test Plan:
Previously the userswitcher only worked on every second try, now it works
always on the first click.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10802