Rename wallpaper translation catalogs to follow pattern
Summary:
The wallpaper loading code expects the catalogs to follow the pattern
"plasma_wallpaper_" + pluginid. While in config QML the domain needs
to be explicitly set using i18nd, as the default textdomain is set to
that of the config container, it is not needed in the wallpaper QML.

While using a catalog with plasma_applet_* prefix works if setting
the domain explicitly, the usage goes against the design of the
naming patterns. So to give a good example and also properly marking
the usage of the catalog, the names are switched.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10850