Let menu change "Items to insert" also create new sections
When switching from LibreOffice splash to e.g. Writer, it creates an entirely new
section for us and then replaces the "root" menu with an alias to that new section.
Since we couldn't create new sections in this fashion we would get an alias to a
non-existing menu and break. This addresses this.