[mediacontroller applet] Use media playback status icons
Summary:
Instead of using the action icons to indicate the current state use
the new media playback status icons.

This is an alternative solution to the problem that the mediacontroller applet
icon in the system tray and the play/pause controll in the player are always
the opposite.

While it is still the opposite icon there is now a slight visual difference
to show that these icons mean different things.

This is meant as a solution to the problem discussed in D10032.

{F5741436}

{F5741437}

Reviewers: #plasma, andreask, broulik

Reviewed By: #plasma, andreask, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11056