KDE logout screen background color fix
Summary:
The patch D5036 did calculate wrong whether the background color is light or not which resulted in always black background. The problem was that color reported by plasma is in range 0.0 -- 1.0 not 0 -- 255. Now the background is black only when button background color is dark and otherwise it respects button background color.

BUG: 382264

Reviewers: mart, #plasma

Reviewed By: mart, #plasma

Subscribers: fredrik, ngraham, abetts, broulik, mvourlakos, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11262