Add a custom default avatar to the sddm-breeze theme.
Summary:
With commit 945a95d908b4544f7c5dea6caacffdc0f44cffcb SDDM will now support a custom default avatar in the next release. [[ https://github.com/sddm/sddm/commit/945a95d908b4544f7c5dea6caacffdc0f44cffcb | Forgot the link. ]]

Before the patch:
{F5741458}
After tha patch:
{F5741459}

Note that the above is not quite the same as the logout screen thingy. But I am not sure how to make it render the image in a way that it does not quite touch the edge of the border.
To illustrate what I mean:
{F5741460}
With the logout screen the user icon is not touching the circle border.

Test Plan: Works for me

Reviewers: #plasma, #vdg, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11057