Pass proper click coordinates to SNI ContextMenu calls
Summary:
Currently it always passes the top left corner of the SNI icon as coordinates,
but now it passes the actual click coordinates.
They can be outside of the icon's bounds (for instance if you click on the
label in the expanded tray), but they are converted to global coordinates
anyway.

Test Plan:
Now the context menu drawn by the application properly follows the
cursor.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11600