Move to KRunner's second results item with a single keypress
Summary:
Currently, to navigate the list of results after typing a search term, you need to press the down arrow twice: first to focus the result list, and a second time to actually change the selection. This patch fixes that by changing the selection during the focus oprtation.

BUG: 392197

Test Plan:
Deployed change, killed, restarted, and activated KRunner, typed a search term:
- The down arrow key now moves to the second item
- Clicked on the text field; up arrow now moves to the previous item in the results list

Activating the {nav Open Containing Folder} button for the first result via the keyboard changes slightly with this patch:
- Currently, it's {nav down arrow > tab key}
- With this patch, it's {nav tab key three times} or {nav down arrow > shift + tab}

Seems like an acceptable change to me to better support the common case of just wanting to easily navigate to the results list items.

Reviewers: #plasma, broulik, mart

Reviewed By: #plasma, broulik, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11611