Vastly simplify image wallpaper code
Summary:
The image wallpaper is a mess of 4 different fade functions, some not in
sync, and very complicated swapping images about.

This simplifies the whole thing by using a StackView, we just pop our
setup with a new configuration on and it handles deleting old objects
when the transition ends.

Code is 1/3 the size, whilst doing the same thing and should result in
having fewer items about.

In addition with this patch changing the transition effect in future
would be super trivial.

Test Plan:
Set as both image and wallpaper
Renders stuff as before

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11726