[ContextMenu Containment Action] Fix checking for KIOSK
Anyone else checks for run_command. Docs explicitly mention the fact that you have to also
disable action/run_command for the menu entry to appear but this is a bit awkward, isn't it.
Also fix logout to check for the correct place as listed in documentation and other code places.

BUG: 393329
FIXED-IN: 5.15.5

Differential Revision: https://phabricator.kde.org/D12376