If GL context creation fails fall back to the software renderer
Summary:
If we cannot create a GL context, instead of showing an error and
quitting, switch to using the QtQuick software renderer and restart.

If software mode is already set, then show the error as before. This
could happen in the case of the software renderer not being installed.

Test Plan:
After doing some driver changing, I ended up with the annoying "plasma can't start"
message.

Without the entry in kdeglobals I now start plasma and it seamlessly comes up
working.

Reviewers: #plasma

Subscribers: mart, broulik, graesslin, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11722