Guard BookmarksRunner thread use
Summary:
Krunner::match is performed in potentially multiple overlapping threads
as a user presses each character. Qt SQL is not thread safe according to
the docs or according to the numerous reports.

It's a bit of a blind patch, ideally we should only use the database in
the thread that opened it.

This patch at least this means queries will be queued which I think will
be enough.

CCBUG: 363136

Test Plan: Restarted krunner. Still saw my firefox bookmarks

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: rempt, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12683