[Notifications] Move "Open With" items into the top group of the menu
Summary:
Move the {nav Open With} items into the top group of the context menu for notifications with images or files that you can open and see in the filesystem. This makes it consistent with Dolphin, which now has all the Open in/with items grouped together at the top of the menu.

I didn't move the {nav Open With} items all the way to the top because my sense is that {nav Open containing folder} is the more common action and deserves to be the top item. Also, putting it below the {nav Open With} actions would result in an extra ugly separator.

Test Plan: {F5848641}

Reviewers: broulik

Reviewed By: broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12882