Systemmonitor: use different label format when the applet is on a vertical panel
Summary:
When the applet is on a vertical panel, its horizontal size is limited, and the speedLabel clashes with the name label. I fixed that by moving the speed values below the sensor name in this case.

I also added some cleanups unintentionally (the first three hunks), as I couldn't find a way to change the diff after it's submitted. They remove an unneeded import, and fix some uninitialized variable warnings. Feel free to ignore them.

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: ngraham, mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12857