[Places Runner] Fix opening devices
The codepath for opening devices was never actually entered as always a QUrl was set as data.
This resulted in errors like "Cannot open URL /org/freedesktop/someudi".
Moreover, for some reason the run() code created a new KFilePlacesModel but since devices are added delayed,
searching for a device in the model right after creation doesn't work.
The code is adjusted to use the one KFilePlacesModel instance in the helper, since run() is executed on the main thread anyway.

Differential Revision: https://phabricator.kde.org/D13840