WIP: libnotificationmanager and new notification applet
libnotificationmanager provides a proper notification model situation similar
to libtaskmanager. There is also a new notification applet implementing the
suggestions and VDG mockups in the wiki

NOTE: the notification dataengine is defunct in this branch, so while the applet
(for now) installs as "new notifications", the old applet will stop working when
you use this. The dataengine will eventually be restored (famous last words)
for compatibility but using the NotificationServer singleton introduced in
libnotificationmanager.

The applet can show notifications, they expire, can be closed, and their
default action invoked. Still heavily WIP and API subject to change.

[1] https://community.kde.org/Plasma/Notifications