Label job notifications with destination file name
If a destination file name is present, it is a better label for a job
than the source file name. This is particularly true for web downloads
where the source file name may be a gibberish base64 string while the
destination file name would be selected by the user.

If there is no destination file name, we fall back to using the source
file name.

BUG: 396744

Differential Revision: https://phabricator.kde.org/D14857