[PanelShadows] Use 0 offset for disabled borders on Wayland
Summary:
On Wayland, if some border is disabled, we don't attach corresponding
empty tile. Yet, we set offset of 1 for that border.

KWin relies on offset values to calculate bounding box for shadows.

For example, if a panel has only top border enabled, then the bounding
box will be 2 pixel wider(and 1px taller) than it should be. This can
result in having 1px thin artefacts on neighboring screens.

In general, we don't have to attach empty tiles for disabled borders.
Thus, we can set corresponding offsets to 0.

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D14999