Allow wallpaper slideshow blur config to persist
Summary:
Slideshow wallpapers use the same Blur/Color selector that single image
wallpapers use.  However, the Blur config entry was added to the image
xml but not the slideshow xml.  The result was that the selection was
never persisted and reverted to the default on restart.  This allows the
slideshow blur config to persist.

Test Plan:
Configure the desktop backgrounds.
Select slideshow with one or more images that are smaller than the screen.
Choose fill with solid color.  Apply and exit.
Note that the background fill takes on the solid color
Restart
Note that the background fill is still the solid color
Configure desktop background
Choose fill with blur.  Apply and exit
Note that the background fill is a blur
Restart
Note that the background fill is still a blur.

Reviewers: #plasma_workspaces, davidedmundson

Reviewed By: #plasma_workspaces, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13555