Merge switch user dialog into lockscreen
Summary:
This reduces a bunch of code, both hidden in the backend as well as the
mostly duplicated front end UI, making it more consistent for users too.

There is a behavioural change that switching user then cancelling will
require your own password.

KSMServer still has the same DBus slot for compatibility which then
proxies over to the screensaver. This could be calling itself, it might
be calling kwin when we're on wayland.

Depends on D15186

Test Plan:
Pressed switch user from the UI
Got a swich user dialog

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: ngraham, mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15187