Move correct indices in the manual sort map when getting a move() call for group children
Summary:
Due to not passing `parent` to QAIM::index(), we were moving top-level
indices in the map instead. This meant the sort map would become out of
sync with the row move, and to the user it would look like both the
group children and unrelated top-level entries moved.

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15550